pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "django-oscar-stubs"
version = "4.1.0b4"
description = "PEP 561 type stubs for django-oscar"
readme = "README.md"
license = { text = "ISC" }
dependencies = [
    "django-oscar>=4.1,<4.2",
    "django-stubs>=5.2.0",
    "django-treebeard-stubs>=4.8.4",
]
requires-python = ">=3.12"

[[project.authors]]
name = "thelab"
email = "thelabdev@thelab.co"

[project.urls]
Homepage = "https://gitlab.com/thelabnyc/stubs/django-oscar-stubs"
Repository = "https://gitlab.com/thelabnyc/stubs/django-oscar-stubs"

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "django-stubs>=5.2.9",
    "pytest>=8.0",
    "pytest-mypy-plugins>=3.0",
    "tox>=4.49.0",
    "tox-uv>=1.33.1",
    "ruff>=0.15.5",
]

[tool.hatch.build.targets.wheel]
packages = ["oscar-stubs", "mypy_oscar_plugin"]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "PYI"]

[tool.ruff.lint.isort]
from-first = true

[tool.commitizen]
name = "cz_conventional_commits"
annotated_tag = true
gpg_sign = false
tag_format = "v$version"
update_changelog_on_bump = true
changelog_merge_prerelease = true
version_provider = "pep621"
version_scheme = "pep440"
version_files = ["pyproject.toml:version"]
pre_bump_hooks = ["pre-commit run --all-files || true", "uv build"]
post_bump_hooks = ["git push origin HEAD $CZ_POST_CURRENT_TAG_VERSION"]
